… api/api-doc/storage_service.json
… dht/token.cc
… docs/dev/topology-over-raft.md
… locator/load_sketch.hh
… locator/tablet_sharder.hh
… locator/tablets.cc
… locator/tablets.hh
… multishard_mutation_query.cc
… mutation_writer/multishard_writer.cc
… mutation_writer/multishard_writer.hh
… readers/multishard.cc
… repair/row_level.cc
… replica/compaction_group.hh
… replica/database.cc
… replica/database.hh
… replica/mutation_dump.cc
… replica/table.cc
… schema/schema.cc
… schema/schema.hh
… scylla-gdb.py
… service/paxos/paxos_state.cc
… service/storage_proxy.cc
… service/storage_service.cc
… service/storage_service.hh
… service/tablet_allocator.cc
… service/topology_coordinator.cc
… sstables/open_info.hh
… sstables/sstable_mutation_reader.cc
… sstables/sstable_mutation_reader.hh
… sstables/sstables.cc