@avikivity 2024-03-26
> Tablet sharder is adjusted to handle intra-migration where a tablet
can have two replicas on the same host. For reads, sharder uses the
read selector to resolve the conflict. For writes, the write selector
is used.

We must drop the use of the sharder for tablets. I'm not proposing to do it now, but the concept was designed for static sharding and doesn't work well for dynamic sharding.